requirements.txt
arrow==1.2.3
asgiref==3.5.2
bleach==4.1.0
blessed==1.19.1
certifi==2022.12.7
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.3
django-filter==22.1
django-jazzmin==2.6.0
django-picklefield==3.1
django-q==1.3.9
django-summernote==0.8.20.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Markdown==3.4.1
packaging==21.3
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
redis==3.5.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
